Move static methods to CloudEnvImpl
